+++ database/migrations/2019_06_17_000010_version_5103.php
@@ -116,6 +116,7 @@ class Version5103 extends Migration
         Setting::saveByKey('purchaseorderEmailSubject', 'Purchase Order #{{ $purchaseorder->number }}');
         Setting::saveByKey('purchaseorderEmailBody', '<p>Please find the attached purchase order from {{ $purchaseorder->user->name }}</p>');
         Setting::saveByKey('skin','{"headBackground":"purple","headClass":"dark","sidebarBackground":"white","sidebarClass":"light","sidebarMode":"open"}');
+        Setting::saveByKey('currencyConversionKey','');
 
         DB::table('schedule_categories')->where('id', 8)->update(['name' => 'Expense and Purchaseorder']);